rename record-notification-manager
