+++ packages/model/src/-private/record-state.ts
@@ -6,7 +6,7 @@ import { cached, tracked } from '@glimmer/tracking';
 import type Store from '@ember-data/store';
 import { storeFor } from '@ember-data/store';
 import { recordIdentifierFor } from '@ember-data/store/-private';
-import type { NotificationType } from '@ember-data/store/-private/managers/record-notification-manager';
+import type { NotificationType } from '@ember-data/store/-private/managers/notification-manager';
 import type RequestCache from '@ember-data/store/-private/network/request-cache';
 import { addToTransaction, subscribe } from '@ember-data/tracking/-private';
 import type { Cache } from '@ember-data/types/q/cache';
