+++ packages/model/src/-private/references/has-many.ts
@@ -13,7 +13,7 @@ import type ManyRelationship from '@ember-data/graph/-private/relationships/stat
 import { DEPRECATE_PROMISE_PROXIES } from '@ember-data/private-build-infra/deprecations';
 import type Store from '@ember-data/store';
 import { recordIdentifierFor } from '@ember-data/store';
-import type { NotificationType } from '@ember-data/store/-private/managers/record-notification-manager';
+import type { NotificationType } from '@ember-data/store/-private/managers/notification-manager';
 import type {
   CollectionResourceDocument,
   CollectionResourceRelationship,
